requirements.txt
langchain
black
python-dotenv
langchain_core
langchain_community